+++ assets/scripts/Plugin/AppCore/_components.ts
@@ -1,5 +1,9 @@
+import ErrorPageButtons from './Error/ErrorPageButtons.vue';
+import ErrorPageImage from './Error/ErrorPageImage.vue';
 import IntroCard from './IntroCard.vue';
 
 export const components = {
+    ErrorPageButtons,
+    ErrorPageImage,
     IntroCard,
 }