+++ src/Authentication/Controller/UserAccountController.php
@@ -6,13 +6,14 @@ namespace App\Authentication\Controller;
 
 use App\Authentication\Entity\Role;
 use App\Core\Controller\AbstractApplicationController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 class UserAccountController extends AbstractApplicationController
 {
-    #[Route('/user/account', name: 'user_account')]
+    #[Route('/user/account', name: 'user_account', methods: [Request::METHOD_GET])]
     #[IsGranted(Role::ROLE_USER)]
     public function dashboard(): Response
     {