+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.2.0] - 2026-04-02
 ### Added
+- Added CSRF protection to all existing forms
 - Added resend verification email button and functionality
 - Added a new `dtmVerified` column to the verification entity 
 - Implemented time limit between token requests